+++ src/analysis/pattern_detection.py
