+++ tests/test_pending_orders.py
@@ -2,14 +2,14 @@ import os
 import sys
 import time
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
+sys.path.append(os.path.join(os.path.dirname(os.path.dirname(__file__)), "src"))
 
 # Configuración común
 os.environ['TAKE_PROFIT_PCT'] = '1.5'
 os.environ['STOP_LOSS_PCT'] = '1'
 
-from bot_trading import FuturesBot, config_por_moneda
-import bot_trading
+from core.bot_trading import FuturesBot, config_por_moneda
+import core.bot_trading as bot_trading
 
 config_por_moneda["TEST/USDT"] = {"atr_factor": 1.0}
